Simkin version 2.18

+++ cpp/Src/skAlist.cpp
@@ -16,7 +16,7 @@
     License along with this library; if not, write to the Free Software
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
-  $Id: skAlist.cpp,v 1.15 2003/02/24 19:59:48 simkin_cvs Exp $
+  $Id: skAlist.cpp,v 1.22 2003/04/11 18:05:39 simkin_cvs Exp $
 */
 #include "skAlist.h"
 
@@ -47,17 +47,20 @@ void skAList::insert(void * p,USize index)
   assert(p);
   assert(index<=m_Entries);
   assert(contains(p)==false);
-  if (index>m_Entries)
+  if (index>m_Entries){
 #ifdef EXCEPTIONS_DEFINED
-		  throw skBoundsException(skSTR("Invalid index to insert"),__FILE__,__LINE__);
+    throw skBoundsException(skSTR("Invalid index to insert"),__FILE__,__LINE__);
 #else
-      exit(EXIT_FAILURE );
+    ExitSystem( );
 #endif
+  }
+  if (m_Array==0)
+    createArray();
   if (m_ArraySize==m_Entries)
     grow();
   if (index<m_Entries)
     for (USize x=m_Entries;x>index;x--)
-      m_Array[x]=m_Array[x-1];		
+      m_Array[x]=m_Array[x-1];          
   m_Array[index]=p;
   m_Entries++;
 }
@@ -67,6 +70,8 @@ void skAList::prepend(void * i)
 {
   assert(i);
   assert(contains(i)==false);
+  if (m_Array==0)
+    createArray();
   if (m_ArraySize==m_Entries)
     grow();
   for (USize x=m_Entries;x>=1;x--)
@@ -77,7 +82,7 @@ void skAList::prepend(void * i)
 //-----------------------------------------------------------------
 void skAList::remove(void * i)
 //-----------------------------------------------------------------
-{   
+{  
   assert(i);
   int index = findElt(i);
   assert(index>=0 && "AList does not contain item"!=0);
@@ -92,28 +97,42 @@ void skAList::removeAndDestroy(void * i)
   assert(i);
   int index = findElt(i);
   assert(index>=0 && "AList does not contain item"!=0);
-
-  deleteItem(m_Array[index]);
-  for (USize x=index;x<m_Entries-1;x++)
-    m_Array[x]=m_Array[x+1];
-
-  m_Entries--;
+  
+  if (m_Array){
+    deleteItem(m_Array[index]);
+    for (USize x=index;x<m_Entries-1;x++)
+      m_Array[x]=m_Array[x+1];
+    m_Entries--;
+  }
 }
 //-----------------------------------------------------------------
 void skAList::deleteElt(USize index)
 //-----------------------------------------------------------------
 {
   assert(index<m_Entries);
-  if (index>=m_Entries)
+  if (index>=m_Entries){
 #ifdef EXCEPTIONS_DEFINED
-		  throw skBoundsException(skSTR("Invalid index to deleteElt"),__FILE__,__LINE__);
+    throw skBoundsException(skSTR("Invalid index to deleteElt"),__FILE__,__LINE__);
 #else
-      exit(EXIT_FAILURE );
+    ExitSystem( );
 #endif
-  deleteItem(m_Array[index]);
-  for (USize x=index;x<m_Entries-1;x++)
-    m_Array[x]=m_Array[x+1];
-  m_Entries--;
+  }else{
+    if (m_Array){
+      deleteItem(m_Array[index]);
+      for (USize x=index;x<m_Entries-1;x++)
+        m_Array[x]=m_Array[x+1];
+      m_Entries--;
+    }
+  }
+}
+//-----------------------------------------------------------------
+void skAList::createArray()
+//-----------------------------------------------------------------
+{
+  if (m_ArraySize==0)
+    m_ArraySize=1;
+  if(m_ArraySize != 0)
+    m_Array=skARRAY_NEW(void *,m_ArraySize);
 }
 //-----------------------------------------------------------------
 void skAList::growTo(USize new_size)
@@ -121,40 +140,35 @@ void skAList::growTo(USize new_size)
 {
   if (new_size>m_ArraySize){
     m_ArraySize = new_size;
-
-    void ** new_array=new void *[m_ArraySize];
     if(m_Array){
+      void ** new_array=skARRAY_NEW(void *,m_ArraySize);
       for (USize x=0;x<m_Entries;x++){
         new_array[x]=m_Array[x];
         m_Array[x]=0;
-	    }
+      }
       delete [] m_Array;
+      m_Array=new_array;
+    }else{
+      m_Array=skARRAY_NEW(void *,m_ArraySize);
     }
-    m_Array=new_array;
   }
 }
 //-----------------------------------------------------------------
 void skAList::grow()
 //-----------------------------------------------------------------
 {
-  if(m_GrowthIncrement != 0)
-    m_ArraySize += m_GrowthIncrement;	// constant increase in size
-  else{
-    if(m_ArraySize==0)
-      m_ArraySize = 1;
-    else
-      m_ArraySize *= 2;			// double size
-  }
-
-  void ** new_array=new void *[m_ArraySize];
+  m_ArraySize *= 2;     // constant increase in size
   if(m_Array){
+    void ** new_array=skARRAY_NEW(void *,m_ArraySize);
     for (USize x=0;x<m_Entries;x++){
       new_array[x]=m_Array[x];
       m_Array[x]=0;
     }
     delete [] m_Array;
+    m_Array=new_array;
+  }else{
+    m_Array=skARRAY_NEW(void *,m_ArraySize);
   }
-  m_Array=new_array;
 }
 //-----------------------------------------------------------------
 void skAList::test() const
